Integrates 43 commits from development branch including:
- hasBattery variable and battery/external power detection (#235)
- Web server improvements (dynamic readings, icons, status page)
- OLED display fixes
- MQTT and preferences refactoring
- Webserver pages overhaul (index, status, preferences, ota)

All modernization/v2 changes preserved:
- ElegantOTA (not AsyncElegantOTA)
- ESP32Async/AsyncTCP + ESPAsyncWebServer
- ArduinoJson 7.0.1 pinned, ArduinoMenu 4.21.4 pinned
- SUPPORT_BLE disabled, -fpermissive added

+++ CO2_Gadget_OLED.h
@@ -35,7 +35,7 @@ void setDisplayReverse(bool reverse) {
     shouldRedrawDisplay = true;
 }
 
-void setDisplayBrightness(uint32_t newBrightness) {
+void setDisplayBrightness(uint16_t newBrightness) {
   Serial.printf("-->[OLED] Setting display brightness value at %d\n", newBrightness);
   u8g2.setContrast(newBrightness);
   actualDisplayBrightness = newBrightness;
